resolvers moved to new file
