+++ services/roomService/src/resolvers/index.ts
@@ -1,20 +1,35 @@
-const { tasks, users } = require('./constants/index.ts');
-const User = require('./data/Models/user.ts')
+const User = require('../data/Models/users/user.model.server.ts') 
 
 module.exports = {
-  Query: {
-    users: () => users,
-    user: (parent: any, args: any) => users.find((user: any) => user.id === args.id)
-  },
+ Query: {
+     user: async (_: any, args: any) => {
+       try {
+         const user = await User.findOne((id: String) => id === args.id)
+         return user;
+       }
+       catch (error) {
+         throw error;
+       }
+     },
+     users: async () => {
+       try {
+         const users = User.find();
+         return users;
+       }
+       catch (error) {
+         throw error;
+       }
+     }
 
+  },
   Mutation: {
-    signUp: async (parent: any, args: any) => {
+    signUp: async (_: any, args: any) => {
       try {
         const user = await User.findOne({ email: args.input.email });
         if (user) {
           throw new Error("User already in use");
         }
-        const newUser = new User({ ...args.input });
+        const newUser = new User({ name: args.input.name, email: args.input.email, password: args.input.password });
         const result = newUser.save();
         return result;
       }
